[Style] 스타일 변경
