+++ webapp/components/PlayController/PlayController.tsx
@@ -1,18 +1,39 @@
-import { ActionIcon, Box, Button, Group, ScrollArea, Stack, Table } from '@mantine/core';
+import {
+  ActionIcon,
+  Box,
+  Button,
+  Checkbox,
+  Group,
+  Modal,
+  ScrollArea,
+  Select,
+  Stack,
+  Table,
+  Text,
+} from '@mantine/core';
+import { showNotification } from '@mantine/notifications';
 import convertActionToNotation from '@store/reversi/convertActionToNotation';
 import useStore from '@store/useStore';
 import { useTranslation } from 'next-i18next';
-import { useCallback, useMemo, useRef } from 'react';
-import { ChevronLeft, ChevronRight, Refresh } from 'tabler-icons-react';
+import { useCallback, useMemo, useRef, useState } from 'react';
+import { Check, ChevronLeft, ChevronRight, Refresh, Settings } from 'tabler-icons-react';
 import shallow from 'zustand/shallow';
+import theme from '~/config/theme';
 
 type PlayControllerProps = {
   onRestart: () => void;
+  onChangeTheme: (value: string) => void;
   isCalculating: boolean;
 };
 
-export default function PlayController({ onRestart, isCalculating }: PlayControllerProps) {
+export default function PlayController({ onRestart, onChangeTheme, isCalculating }: PlayControllerProps) {
   const { t } = useTranslation('common');
+
+  // 모달 관련
+  const themeNames = Object.keys(theme);
+  const [selectedTheme, setSelectedTheme] = useState(themeNames[0]);
+  const [checkedSoundOn, setCheckedSoundOn] = useState(true);
+  const [openedSettingModal, setOpenedSettingModal] = useState(false);
 
   const viewportRef = useRef<HTMLDivElement>(null);
   const scrollToBottom = useCallback(() => {
@@ -34,10 +55,14 @@ export default function PlayController({ onRestart, isCalculating }: PlayControl
     // 두 개씩 묶어서 테이블 형태로 보여줌
     for (let i = 0; i < histories.length; i += 2) {
       result.push(
-        <tr key={i}>
-          <td style={{ width: 50 }}>{`${i / 2 + 1}.`}</td>
-          <td align="center">{convertActionToNotation(histories.at(i)?.action)}</td>
-          <td align="center">{convertActionToNotation(histories.at(i + 1)?.action)}</td>
+        <tr key={i} style={{}}>
+          <td align="center" style={{ verticalAlign: 'bottom' }}>{`${i / 2 + 1}.`}</td>
+          <td align="center" style={{ verticalAlign: 'bottom' }}>
+            <Text weight={500}>{convertActionToNotation(histories.at(i)?.action)}</Text>
+          </td>
+          <td align="center" style={{ verticalAlign: 'bottom' }}>
+            <Text weight={500}>{convertActionToNotation(histories.at(i + 1)?.action)}</Text>
+          </td>
         </tr>,
       );
     }
@@ -54,24 +79,69 @@ export default function PlayController({ onRestart, isCalculating }: PlayControl
         width: 280,
       })}
     >
-      <Stack>
+      <Stack justify="space-between" sx={{ height: '100%' }}>
         <ScrollArea style={{ height: 300 }} viewportRef={viewportRef}>
-          <Table striped highlightOnHover>
+          <Table striped>
             <tbody>{historiesData}</tbody>
           </Table>
         </ScrollArea>
-        <Group grow spacing="xs">
-          <ActionIcon variant="outline" size="lg" onClick={() => {}} disabled={isCalculating}>
+        <Group grow spacing="xs" position="right">
+          <ActionIcon variant="light" size="xl" onClick={() => setOpenedSettingModal(true)}>
+            <Settings size={18} />
+          </ActionIcon>
+          <ActionIcon variant="light" size="xl" onClick={() => {}} disabled={isCalculating}>
             <ChevronLeft size={18} />
           </ActionIcon>
-          <ActionIcon variant="outline" size="lg" onClick={() => {}} disabled={isCalculating}>
+          <ActionIcon variant="light" size="xl" onClick={() => {}} disabled={isCalculating}>
             <ChevronRight size={18} />
           </ActionIcon>
-          <ActionIcon variant="outline" size="lg" onClick={onRestart} disabled={isCalculating}>
-            <Refresh size={18} />
+          <ActionIcon variant="light" size="xl" onClick={onRestart} disabled={isCalculating}>
+            <Refresh size={16} />
           </ActionIcon>
         </Group>
       </Stack>
+      <Modal
+        opened={openedSettingModal}
+        centered
+        onClose={() => setOpenedSettingModal(false)}
+        title="게임 설정"
+      >
+        <Stack>
+          <Select
+            label="테마"
+            value={selectedTheme}
+            data={themeNames}
+            onChange={(value) => {
+              if (!value) return setSelectedTheme(themeNames[0]);
+              setSelectedTheme(value);
+              onChangeTheme(value);
+            }}
+          />
+          <Checkbox
+            label="소리 재생"
+            checked={checkedSoundOn}
+            onChange={(e) => setCheckedSoundOn(e.currentTarget.checked)}
+          />
+          <Group position="right">
+            <Button variant="default" onClick={() => setOpenedSettingModal(false)}>
+              취소
+            </Button>
+            <Button
+              onClick={() => {
+                setOpenedSettingModal(false);
+                showNotification({
+                  title: '설정 저장 완료',
+                  message: '설정이 저장되었습니다.',
+                  color: 'blue',
+                  icon: <Check size={16} />,
+                });
+              }}
+            >
+              저장
+            </Button>
+          </Group>
+        </Stack>
+      </Modal>
     </Box>
   );
 }
