+++ webapp/layouts/AppNavBar/index.tsx
@@ -5,7 +5,7 @@ import { Navbar } from '@mantine/core';
 export default function AppNavBar({ hidden }: { hidden: boolean }) {
   return (
     <Navbar width={{ sm: 200, lg: 300 }} p="xs" hiddenBreakpoint="sm" hidden={!hidden}>
-      <Navbar.Section grow mt="xs">
+      <Navbar.Section grow>
         <MainLinks />
       </Navbar.Section>
       <Navbar.Section>
