+++ webapp/public/locales/ko/navbar.json
@@ -1,7 +1,7 @@
 {
     "play": "플레이",
     "community": "커뮤니티",
-    "my-games": "내 게임",
+    "my-games": "내 전적",
     "guest": "게스트",
     "application": "애플리케이션",
     "login": "로그인",
